[2주차] App 컴포넌트 분리
- index.jsx는 ReactDOM 역할을 하는 파일로 두고 APP 컴포넌트를 따로
분리하여 관리한다
